Add Games as type for widgets
In my overrides.xml, I was unable to set a widget with type "games", as there was no string type that pointed to games. There were only strings 32009 = Program, 32010 = Video Add-on. 32011 = Music Add-on and 32012 = Picture Addon. So, I added a new string (32123 = Games). So now the game widgets open in the Games.xml window.